api/telemetric_system/requirements/base.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis==5.2.0

# MQTT
paho-mqtt==2.1.0

# Security
bcrypt==4.2.1
pyjwt==2.10.1
python-multipart==0.0.20

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# HTTP client
requests==2.32.3

# Data processing
pandas==2.2.3
numpy==2.2.1

# Geospatial
geopy==2.4.1

# Background tasks (optional)
celery==5.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
